github: add actions to automate backporting fixes
This adds support for automatically cherry-picking fixes for the
release branch using a special 'command' in issue comments:

/cherry-pick <commit1> <commit2> ...

This command attempts to cherry-pick the specified commit(s) to the
release branch based on the milestone attached to the issue or pull
request. If the commit(s) fail to apply cleanly, then a comment with
a link to the failing job will be added to the issue/pull request.
If the commit(s) do apply cleanly, then a pull request will be
created with the specified commits.